refactor iOS code
