+++ packages/eas-cli/src/run/ios/simctl.ts
@@ -1,0 +1,10 @@
+import { SpawnOptions, SpawnResult } from '@expo/spawn-async';
+
+import { xcrunAsync } from './xcrun';
+
+export async function simctlAsync(
+  args: (string | undefined)[],
+  options?: SpawnOptions
+): Promise<SpawnResult> {
+  return xcrunAsync(['simctl', ...args], options);
+}
